refactor detail page: info and menus
- add new sub component, menus and info
- uncomment the index.js and webpack.common.js
