+++ src/views/detail-view/detail-view.scss
@@ -26,103 +26,6 @@ detail-view {
         gap: 3rem;
     }
 
-    .detail {
-        &__restaurant-name {
-            font-size: 1.75rem;
-            max-width: 1280px;
-            margin-right: 1.5rem;
-            margin-bottom: 0.5rem;
-            display: inline-block;
-        }
-
-        &__fav-button {
-            @include flex-center(row);
-            position: fixed;
-            bottom: 2rem;
-            right: 2rem;
-            cursor: pointer;
-            border-radius: 50%;
-            font-size: 1.5rem;
-            height: 44px;
-            width: 44px;
-            border: none;
-            background: $color-rating-star;
-            
-            i.favorite {
-                color: $white;
-            }
-    
-            @include responsive-min($medium) {
-                position: relative;
-                display: inline;
-                bottom: initial;
-                right: initial;
-            }
-        }
-
-        &__address {
-            margin: 0.25rem 0;
-            color: $color-text-subtle;
-            text-transform: capitalize;
-        }
-
-        &__categories {
-            margin: 1rem 0;
-        }
-
-        &__category {
-            border-radius: 0.25rem;
-            padding: 0.25rem 0.5rem;
-            background: $color-rating-star;
-        }
-
-        &__desc {
-            margin-top: 2rem;
-            overflow: hidden;
-            line-height: 1.6;
-            text-overflow: ellipsis; 
-            -webkit-box-orient: vertical; 
-            display: -webkit-box; 
-            
-            &.clipped {
-                -webkit-line-clamp: 3;
-            }
-        }
-
-        &__toggle-desc {
-            border: none;
-            background: none;
-            color: $dark-orange;
-            font-size: 1rem;
-            padding: 0.25rem 0;
-            min-height: 44px;
-            cursor: pointer;
-        }
-    }
-
-    .menu {
-        &__wrapper {
-            display: grid;
-            gap: 1rem;
-        }
-
-        &__category {
-            text-align: center;
-            text-transform: capitalize;
-            padding: 1.5rem;
-        }
-
-        &__item {
-            display: block;
-            min-height: 44px;
-            padding: 0.75rem 0;
-            text-align: center;
-            list-style-type: none;
-            text-transform: capitalize;
-            border-bottom: 0.5px solid $color-text-subtle;
-        }
-    }
-
     .reviews {
         &__form {
             margin: 2rem 0;
