+++ src/views/detail-view/detail-view_responsive.scss
@@ -1,21 +1,6 @@
 @import 'src/styles/config';
 
 detail-view {
-    .menu {
-        &__wrapper {
-            @include responsive-min($medium) {
-                grid-template-columns: repeat(2, 1fr);
-            }
-        }
-    
-        &__title {
-            @include responsive-min($medium) {
-                text-align: center;
-            }
-        }
-    }
-
-
     .reviews {
         &__title {
             @include responsive-min($medium) {
