+++ src/views/detail-view/restaurant-menus/restaurant-menus.scss
@@ -1,0 +1,31 @@
+@import 'src/styles/config';
+
+restaurant-menus {
+    .menu {
+        &__wrapper {
+            display: grid;
+            gap: 1rem;
+        }
+
+        &__category {
+            text-align: center;
+            text-transform: capitalize;
+            padding: 1.5rem;
+        }
+
+        &__item {
+            display: block;
+            min-height: 44px;
+            padding: 0.75rem 0;
+            text-align: center;
+            list-style-type: none;
+            text-transform: capitalize;
+            border-bottom: 0.5px solid $color-text-subtle;
+        }
+    }
+
+    .divider {
+        border: 0.5px solid $color-rating-star;
+        margin: 0.5rem 0;
+    }
+}
